minedeep_crawler/requirements.txt
matplotlib==2.1.2
numpy==1.13.3
tweepy==3.6.0
pandas==0.22.0
nltk==3.2.5
python_dateutil==2.7.3
mysql_connector_repackaged==0.3.1
scikit_learn==0.19.1
